+++ Common/src/main/java/fuzs/easyshulkerboxes/api/config/ServerConfigCore.java
@@ -3,4 +3,6 @@ package fuzs.easyshulkerboxes.api.config;
 public interface ServerConfigCore {
 
     boolean allowSlotCycling();
+
+    boolean allowMouseDragging();
 }