+++ Common/src/main/java/fuzs/easyshulkerboxes/world/inventory/ShulkerBoxProvider.java
@@ -4,9 +4,6 @@ import fuzs.easyshulkerboxes.EasyShulkerBoxes;
 import fuzs.easyshulkerboxes.api.world.inventory.ContainerItemProvider;
 import fuzs.easyshulkerboxes.api.world.item.container.ContainerItemHelper;
 import fuzs.easyshulkerboxes.config.ServerConfig;
-import fuzs.puzzleslib.proxy.Proxy;
-import net.minecraft.world.SimpleContainer;
-import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.tooltip.TooltipComponent;
 import net.minecraft.world.item.DyeColor;
 import net.minecraft.world.item.ItemStack;
@@ -15,31 +12,16 @@ import net.minecraft.world.level.block.entity.BlockEntityType;
 
 import java.util.Optional;
 
-public class ShulkerBoxProvider extends ContainerItemProvider {
+public class ShulkerBoxProvider extends ItemWithBlockEntityProvider {
     public static final ContainerItemProvider INSTANCE = new ShulkerBoxProvider();
 
     private ShulkerBoxProvider() {
-
-    }
-
-    @Override
-    public SimpleContainer getItemContainer(Player player, ItemStack stack, boolean allowSaving) {
-        return ContainerItemHelper.loadItemContainer(stack, BlockEntityType.SHULKER_BOX, 27, allowSaving);
+        super(BlockEntityType.SHULKER_BOX, 9, 3);
     }
 
     @Override
     protected boolean canItemFitInside(ItemStack containerStack, ItemStack stack) {
         return stack.getItem().canFitInsideContainerItems();
-    }
-
-    @Override
-    protected boolean _canAcceptItem(ItemStack containerStack, ItemStack stack) {
-        return this.getItemContainer(Proxy.INSTANCE.getClientPlayer(), containerStack, false).canAddItem(stack);
-    }
-
-    @Override
-    protected int _getAcceptableItemCount(ItemStack containerStack, ItemStack stack) {
-        return stack.getCount();
     }
 
     @Override
@@ -51,10 +33,10 @@ public class ShulkerBoxProvider extends ContainerItemProvider {
     public Optional<TooltipComponent> getTooltipImage(ItemStack stack) {
         DyeColor color = ShulkerBoxBlock.getColorFromItem(stack.getItem());
         if (color != null) {
-            return ContainerItemHelper.getTooltipImage(stack, BlockEntityType.SHULKER_BOX, 3, color);
+            return ContainerItemHelper.getTooltipImage(stack, this.blockEntityType, this.inventoryWidth, this.inventoryHeight, color);
         } else {
             // beautiful lavender color from Tinted mod once again
-            return ContainerItemHelper.getTooltipImageWithColor(ContainerItemHelper.getTooltipContainer(stack, BlockEntityType.SHULKER_BOX, 27), 3, new float[]{0.88235295F, 0.6901961F, 0.99607843F});
+            return ContainerItemHelper.getTooltipImageRaw(ContainerItemHelper.getTooltipContainer(stack, this.blockEntityType, this.getInventorySize()), this.inventoryWidth, this.inventoryHeight, new float[]{0.88235295F, 0.6901961F, 0.99607843F});
         }
     }
 }