+++ Fabric/src/main/java/fuzs/easyshulkerboxes/mixin/client/MouseHandlerMixin$Fabric.java
@@ -14,8 +14,9 @@ abstract class MouseHandlerMixin$Fabric {
     @Shadow
     private int activeButton;
 
-    @Inject(method = "method_1602(Lnet/minecraft/client/gui/screens/Screen;DDDD)V", at = @At("HEAD"), remap = false, cancellable = true)
+    @Inject(method = "method_1602(Lnet/minecraft/client/gui/screens/Screen;DDDD)V", at = @At("HEAD"), cancellable = true)
     private void simpleinventorycontainers$onMove(Screen screen, double mouseX, double mouseY, double dragX, double dragY, CallbackInfo callback) {
+        // we need to take over everything to be able to capture the result from Screen::mouseDragged
         callback.cancel();
         if (MouseDragEvents.BEFORE.invoker().beforeMouseDrag(screen, mouseX, mouseY, this.activeButton, dragX, dragY).isPresent()) return;
         if (screen.mouseDragged(mouseX, mouseY, this.activeButton, dragX, dragY)) return;