+++ Forge/src/main/java/fuzs/easyshulkerboxes/client/EasyShulkerBoxesForgeClient.java
@@ -8,9 +8,9 @@ import fuzs.easyshulkerboxes.config.ClientConfig;
 import fuzs.easyshulkerboxes.config.ServerConfig;
 import fuzs.puzzleslib.client.core.ClientFactories;
 import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.client.event.ContainerScreenEvent;
 import net.minecraftforge.client.event.ScreenEvent;
 import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.event.PlayLevelSoundEvent;
 import net.minecraftforge.event.entity.EntityJoinLevelEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
@@ -33,7 +33,7 @@ public class EasyShulkerBoxesForgeClient {
             MouseScrollHandler.onMouseScroll(evt.getScreen(), evt.getMouseX(), evt.getMouseY(), evt.getScrollDelta(), evt.getScrollDelta(), EasyShulkerBoxes.CONFIG.get(ClientConfig.class), EasyShulkerBoxes.CONFIG.get(ServerConfig.class)).ifPresent(unit -> evt.setCanceled(true));
         });
         MinecraftForge.EVENT_BUS.addListener((final ScreenEvent.MouseButtonPressed.Pre evt) -> {
-            MouseDragHandler.INSTANCE.onMousePress(evt.getScreen(), evt.getMouseX(), evt.getMouseY(), evt.getButton()).ifPresent(unit -> evt.setCanceled(true));
+            MouseDragHandler.INSTANCE.onMousePress(evt.getScreen(), evt.getMouseX(), evt.getMouseY(), evt.getButton(), EasyShulkerBoxes.CONFIG.get(ServerConfig.class)).ifPresent(unit -> evt.setCanceled(true));
         });
         MinecraftForge.EVENT_BUS.addListener((final ScreenEvent.MouseDragged.Pre evt) -> {
             MouseDragHandler.INSTANCE.onMouseDrag(evt.getScreen(), evt.getMouseX(), evt.getMouseY(), evt.getMouseButton(), evt.getDragX(), evt.getDragY()).ifPresent(unit -> evt.setCanceled(true));
@@ -41,11 +41,8 @@ public class EasyShulkerBoxesForgeClient {
         MinecraftForge.EVENT_BUS.addListener((final ScreenEvent.MouseButtonReleased.Pre evt) -> {
             MouseDragHandler.INSTANCE.onMouseRelease(evt.getScreen(), evt.getMouseX(), evt.getMouseY(), evt.getButton()).ifPresent(unit -> evt.setCanceled(true));
         });
-        MinecraftForge.EVENT_BUS.addListener((final ContainerScreenEvent.Render.Foreground evt) -> {
-            MouseDragHandler.INSTANCE.onDrawForeground(evt.getContainerScreen(), evt.getPoseStack(), evt.getMouseX(), evt.getMouseY());
+        MinecraftForge.EVENT_BUS.addListener((final PlayLevelSoundEvent.AtEntity evt) -> {
+            MouseDragHandler.INSTANCE.onPlaySoundAtPosition(evt.getEntity(), evt.getSound(), evt.getSource(), evt.getOriginalVolume(), evt.getOriginalPitch()).ifPresent(unit -> evt.setCanceled(true));
         });
-//        MinecraftForge.EVENT_BUS.addListener((final TickEvent.ClientTickEvent evt) -> {
-//            if (evt.phase == TickEvent.Phase.END) MouseScrollHandler.onClientTick$End(Minecraft.getInstance(), EasyShulkerBoxes.CONFIG.get(ClientConfig.class));
-//        });
     }
 }